refactor repo - mgfn
