+++ scripts/convert_official_to_hf.py
