+++ src/configuration_mgfn.py
