+++ src/i3d.py
@@ -1,6 +1,3 @@
-import os
-import math
-
 from typing import Callable, Tuple
 
 import torch
